requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.1
Flask==3.0.3
requests==2.31.0
mysql-connector-python==8.4.0
